Change display of category pages to show published and draft content next to each other

+++ src/app/code/community/Firegento/FlexCms/Model/Observer.php
@@ -161,7 +161,9 @@ class Firegento_FlexCms_Model_Observer
             if ($category->getId()) {
                 /** @var $changesObject Firegento_FlexCms_Model_Category_Changes */
                 $changesObject = Mage::getModel('firegento_flexcms/category_changes')->loadByCategory($category);
-                $changesObject->delete();
+                if ($changesObject->getId()) {
+                    $changesObject->delete();
+                }
             }
             return;
         }
